// dist/node/pkc/pkc.d.ts
export declare class PKC extends PKCTypedEmitter<PKCEvents> implements ParsedPKCOptions {
    ipfsGatewayUrls: ParsedPKCOptions["ipfsGatewayUrls"];
    kuboRpcClientsOptions?: ParsedPKCOptions["kuboRpcClientsOptions"];
// ...
}